Fix printing of "reduced by" in terminal output
Fixed and cleaned all integer-to-integer divisions.

+++ src/neper_v/nev_print/nev_print_mesh/nev_print_mesh_0d/nev_print_mesh_0d1.c
@@ -47,7 +47,7 @@ nev_print_mesh_0d (FILE * file, struct PRINT Print, struct MESH *Mesh,
                       "Number of 0D elts      reduced by %3.0f%% (to %d).\n",
                       100 * (double) (MeshData[0].EltQty -
                                       printelt0d_qty) /
-                      (double) MeshData[0].EltQty, printelt0d_qty);
+                      MeshData[0].EltQty, printelt0d_qty);
   }
 
   return;
