+++ src/neper_v/nev_print/nev_print_mesh/nev_print_mesh_1d/nev_print_mesh_1d1.c
@@ -108,7 +108,7 @@ nev_print_mesh_1d (FILE * file, struct PRINT Print, struct TESS Tess,
 
   ut_print_message (0, 4,
                     "Number of 1D elts      reduced by %3.0f%% (to %d).\n",
-                    100 * (double) hidden[0] / (double) elt1d_qty,
+                    100 * (double) hidden[0] / elt1d_qty,
                     elt1d_qty - hidden[0]);
 
   ut_free_1d_int (&hidden);
